requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.1